+++ components/avatar/avatar.tsx
@@ -7,8 +7,8 @@ import { responsiveArray } from '../_util/responsiveObserver';
 import warning from '../_util/warning';
 import { ConfigContext } from '../config-provider';
 import useBreakpoint from '../grid/hooks/useBreakpoint';
-import type { AvatarContextType, AvatarSize } from './SizeContext';
-import SizeContext from './SizeContext';
+import type { AvatarContextType, AvatarSize } from './avatarContext';
+import SizeContext from './avatarContext';
 import useStyle from './style';
 
 export interface AvatarProps {
